Ref: generic emission - another approach.
